requirements.txt
# Core ML and Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.3
torch==2.1.0
transformers==4.35.2

# API and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Kafka and Streaming
kafka-python==2.0.2
confluent-kafka==2.3.0

# Monitoring and Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Database and Storage
redis==5.0.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
joblib==1.3.2
pyyaml==6.0.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0